pyproject.toml
[project]
name = "a2a-lab"
version = "0.1.0"
description = "Agent-to-Agent (A2A) Protocol Lab for Applied Generative AI course"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "strands-agents[a2a]>=0.1.6",
    "litellm>=1.30.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "aiohttp>=3.8.0",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "ipykernel>=6.29.5",
    "jupyter-client>=8.6.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = false
